Making component function to instantiate components
